requirements.txt
tempfile
pillow
av
mediapipe